Use off_t rather than double where possible
Beautiful of libcurl to use floating point types for what are never
fractional values. We can do better, and we usually want these values in
their integer form anyway.

+++ lib/libalpm/dload.h
@@ -33,8 +33,8 @@ struct dload_payload {
 	char *destfile_name;
 	char *content_disp_name;
 	char *fileurl;
-	double initial_size;
-	long max_size;
+	off_t initial_size;
+	off_t max_size;
 	int force;
 	int allow_resume;
 	int errors_ok;
